refactor(app): improve code quality and readability
- Reduce visibility of methods and fields
- Use final modifiers where appropriate
- Simplify code in several classes
- Remove unnecessary type casts
- Use more descriptive variable names

+++ app/src/main/java/dx/rop/cst/CstLiteral32.java
@@ -82,6 +82,6 @@ public abstract class CstLiteral32
     /** {@inheritDoc} */
     @Override
     public final long getLongBits() {
-        return (long) bits;
+        return bits;
     }
 }
